double型以外の引数の型チェックと戻り値を実装

+++ mogral/src/code_gen/error.rs
@@ -14,7 +14,7 @@ pub enum CodeGenError {
     #[error("variable already exists: `{0}`")]
     VariableAlreadyExists(String),
     #[error("invalid number of arguments passed (expected {expected}, found {found})")]
-    InvalidNumberOfArguments { expected: u32, found: u32 },
+    InvalidNumberOfArguments { expected: usize, found: usize },
     #[error("mismatched types (expected {expected}, found {found})")]
     MismatchedTypes { expected: MglType, found: MglType },
     #[error("invalid operand types ({lhs} {op} {rhs})")]
